+++ components/brave_vpn/browser/connection/wireguard/win/brave_vpn_wireguard_service/brave_vpn_wireguard_service.rc
@@ -1,4 +1,4 @@
 #include "brave/components/brave_vpn/browser/connection/wireguard/win/brave_vpn_wireguard_service/status_tray/resources/resources_en.rc"
 
-1 TYPELIB "brave/components/brave_vpn/browser/connection/wireguard/win/brave_vpn_wireguard_service/service/brave_wireguard_manager_idl.tlb"
+1 TYPELIB "brave/components/brave_vpn/common/wireguard/win/brave_wireguard_manager_idl.tlb"
