+++ components/brave_vpn/common/wireguard/win/BUILD.gn
@@ -3,9 +3,18 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at https://mozilla.org/MPL/2.0/.
 
+import("//brave/build/config.gni")
 import("//brave/components/brave_vpn/common/buildflags/buildflags.gni")
+import("//build/toolchain/win/midl.gni")
 
+assert(is_win)
 assert(enable_brave_vpn)
+
+midl("brave_wireguard_manager_idl") {
+  sources = [ "brave_wireguard_manager_idl.idl" ]
+
+  writes_tlb = true
+}
 
 source_set("win") {
   sources = [
@@ -15,8 +24,8 @@ source_set("win") {
     "wireguard_utils.h",
   ]
   deps = [
+    ":brave_wireguard_manager_idl",
     "//base",
-    "//brave/components/brave_vpn/browser/connection/wireguard/win/brave_vpn_wireguard_service/service:brave_wireguard_manager_idl",
     "//brave/components/brave_vpn/common/buildflags",
     "//brave/components/brave_vpn/common/win",
     "//brave/components/brave_vpn/common/wireguard",
