+++ components/brave_vpn/common/wireguard/win/brave_wireguard_manager_idl.idl
