+++ win_build_output/midl/brave/components/brave_vpn/common/wireguard/win/x64/brave_wireguard_manager_idl.tlb
