+++ win_build_output/midl/brave/components/brave_vpn/common/wireguard/win/x86/brave_wireguard_manager_idl.dlldata.c
